Replace `requestFocus()` by `requestFocusInWindow()` where applicable
The former gives focus to the window in which a component is present,
while the latter only changes the focus within the current window (not
focusing the window itself if it is not focused yet). Java documentation
recommends changing `requestFocusInWindow()` where possible, due to some
platform-dependent behaviour in `requestFocus()`.

When focusing the serial monitor and plotter, `requestFocus()` is still
used, since then the focused window *should* change.