Remove `Editor.handleStop()`
It did not really do anything, other than clearing the active state on
the "verify" toolbar button and bringing the IDE to the front. This code
was only run on save and save as, the stop button has been gone for a
while it seems. Proper support for stopping compilations and upload
should be added, but this code has no value for when that happens.