Replace most SimpleAction constructors with chainable setters
Before, there was a constructor for every combination of properties, which
resulted in 8 constructors already and would explode when multiple
icon properties will be added later. With this change, each property has
its own chainable setter, allowing things like:

    public final Action prevTab = new SimpleAction(tr("Previous Tab"))
       .accelerator(Keys.ctrlAlt(KeyEvent.VK_LEFT))
       .listener(() -> editor.selectPrevTab());

This has the advantage that any combination is supported, and also that
each property value is annoted by the setter method used, making it
easier to read.