Add "selected" and "rollover" icon properties to SimpleAction
Normally, the Java Action class only defines a small and large icon.
However, JButtons/AbstractButtons can also have a "selected" icon, used
when a button is selected (used for e.g. JRadioButton), and a "rollover"
icon, when the mouse is hovering over the button. Java itself does not
offer any builtin way to set these through an action, so this adds two
new properties for that, along with setters for them.

This just prepares the SimpleAction class for the use of these
properties, it does not actually let any button classes make use of them
yet.