Use Action objects for file, sketch and tools menu
Instead of defining various properties (name, accelerator, handling) of
these menu items when creating them, this defines Action objects
containing these properties instead. When creating the menu, only a
reference to the Action object is passed to JMenuItem.

A big advantage of this is that when the same action is available from
mulitiple places (e.g. menu and toolbar), all places can refer to the
same action without duplicating properties (though this commit does not
update anything but the menus yet). Since an Action can additionally
store "enabled" and "selected" flags, these will be automatically shared
between all menu items, buttons, etc. using this action, making the
bookkeeping of where to enable/disable items a lot easier (though this
commit does not apply this yet, menu items are still enabled/disabled
directly).

The actions are declared as public final attributes of an Editor
instance, where they are easily accessible to any code that needs them.
One downside of this approach is that, when they are created, the
Editor constructor has not run yet. This means that when the action, and
also the lambda handler for the action is created, any instance
variables will not have been initialized yet. Once the lambda handler
actually runs, this has long been resolved, but the compiler does not
know this and thus complains about referencing potentially uninitialized
variables. For this reason, a few additional `handleXXX()` methods have
been added, which usually do nothing more than call a method on some
instance variable (often `base`).

Ideally, more of the handling of these methods would be moved into the
lambda, but then the way of instantiating these Action objects might
need to be changed. Alternatively, this could be made more consistent by
creating a `handleXXX()` method for each action (but the ones that do
not already call existing `handleXXX()` methods that probably need a bit
of cleanup anyway, so these were left as is for now).

The sketch menu is left as it is, since those actions should probably
end up in EditorTab, so that needs some more thought. The help menu has
been left alone as well, since that is fairly self-contained. It might
need to be converted later as well, though.