Add icons to all toolbar actions
This adds icons for all actions that are present on the toolbar and sets
them on the corresponding Action objects. These icons are created by
slicing up buttons.gif, which contains all of these icons in a single
image.

These icons are not currently used yet, but this commit prepares for
EditorToolbar making use of the Action objects in a subsequent commit.