WIP: Custom formatters
This allows passing formatters to print. E.g. instead of print(10, HEX),
you can pass any object that has a suitable printTo(Print*, ValueT)
method.

This is a WIP on top of the variadic print PR and it makes some changes
that should probably go into there.

As an example of usage:

	struct EvenOddFormatter {
	  int printTo(Print *p, int x) {
	    return p->print(x % 2 ? "odd" : "even");
	  }
	};

	Serial.println(1, EvenOddFormatter());