Store original board items, not clones and use the original action
This simplifies the creation of the original board items by just storing
the original and leaving the creation of new actions and items for the
recent boards menu for later, for only the items actually present there.

Also, it turns out that cloning the action is not actually needed, you
can just attach the original action to the recent board menu item as
well (this skips setting the clicked item as selected, but it works
because the original action changes the current board, which regenerates
the recent board menu anyway).